refactor: extract DatabasePartition from database.cpp
